Procedulary Generated Scene Update

+++ src/Additional/GameScripts/include/PlayerMovement.h
@@ -53,7 +53,7 @@ class PlayerMovement : public Component {
 		bool reachEnd = true;
 		glm::vec3 destination;
 		float speed = 0.7f;
-		int maxSteps = 5;
+		size_t maxSteps = 5;
 		float nextWaypointDistance = 0.05f;
 		
 
@@ -61,7 +61,7 @@ class PlayerMovement : public Component {
 
 		LineRenderer* circleRenderer;
 		float radius = 3.0f;
-		int steps = 20;
+		size_t steps = 20;
 
 		//Line
